requirements.txt
plotly
scipy
seaborn
matplotlib
sklearn
PyPDF2
nltk
bokeh
streamlit-bokeh-events
mysql-connector-python
